requirements.txt
langchain_community
langchain_ollama
langchain_chroma
langchain
chromadb
pypdf
fastembed
discord.py